added transfer-request system

+++ wallet/models.py
@@ -5,7 +5,7 @@ from account.models import Currency
 
 class Wallet(models.Model):
     user=models.OneToOneField(User,on_delete=models.CASCADE,related_name='wallet')
-    balance=models.IntegerField(default=1000)
+    balance=models.DecimalField(default=1000,decimal_places=2,max_digits=10)
     currency=models.CharField(max_length=5,choices=Currency.choices)
 
     def __str__(self) -> str:
